rename modules root paths
