+++ core/audit/src/main/java/ru/arooid/financial/audit/Application.java
@@ -1,4 +1,4 @@
-package ru.arooid.financial.gateway;
+package ru.arooid.financial.audit;
 
 import io.micronaut.runtime.Micronaut;
